Create Clear First Stage Recycle bin of Site collection in SharePoint Online
Script Overview:
This script is designed to clear the first-stage recycle bin for a list of site collections in SharePoint Online, which is stored in a CSV file.

To run this script smoothly, set up an application in your Azure AD through App registration. Follow the steps in the article:
https://learn.microsoft.com/en-us/entra/identity-platform/quickstart-register-app

For authentication to SharePoint Online using this app, it's recommended to use the certificate method for better security. Refer to the article by LAURA KOKKARINEN for detailed steps:
https://laurakokkarinen.com/authenticating-to-office-365-apis-with-a-certificate-step-by-step/